Detect connect/disconnect monitor

+++ SnapIt.Library/Extensions/ScreenExtensions.cs
@@ -1,8 +1,9 @@
 ﻿using System;
 using System.Drawing;
 using System.Runtime.InteropServices;
-using System.Windows.Forms;
+
 using SnapIt.Library.Entities;
+using WpfScreenHelper;
 
 namespace SnapIt.Library.Extensions
 {
@@ -21,29 +22,29 @@ namespace SnapIt.Library.Extensions
 
         //https://msdn.microsoft.com/en-us/library/windows/desktop/dd145062(v=vs.85).aspx
         [DllImport("User32.dll")]
-        private static extern IntPtr MonitorFromPoint([In]Point pt, [In]uint dwFlags);
+        private static extern IntPtr MonitorFromPoint([In] Point pt, [In] uint dwFlags);
 
         //https://msdn.microsoft.com/en-us/library/windows/desktop/dn280510(v=vs.85).aspx
         [DllImport("Shcore.dll")]
-        private static extern IntPtr GetDpiForMonitor([In]IntPtr hmonitor, [In]DpiType dpiType, [Out]out uint dpiX, [Out]out uint dpiY);
+        private static extern IntPtr GetDpiForMonitor([In] IntPtr hmonitor, [In] DpiType dpiType, [Out] out uint dpiX, [Out] out uint dpiY);
     }
 
     public static class ScreenExtensions
     {
         public static void GetDpi(this Screen screen, DpiType dpiType, out uint dpiX, out uint dpiY)
         {
-            var pnt = new Point(screen.Bounds.Left + 1, screen.Bounds.Top + 1);
+            var pnt = new Point((int)screen.Bounds.Left + 1, (int)screen.Bounds.Top + 1);
             var mon = MonitorFromPoint(pnt, 2/*MONITOR_DEFAULTTONEAREST*/);
             GetDpiForMonitor(mon, dpiType, out dpiX, out dpiY);
         }
 
         //https://msdn.microsoft.com/en-us/library/windows/desktop/dd145062(v=vs.85).aspx
         [DllImport("User32.dll")]
-        private static extern IntPtr MonitorFromPoint([In]Point pt, [In]uint dwFlags);
+        private static extern IntPtr MonitorFromPoint([In] Point pt, [In] uint dwFlags);
 
         //https://msdn.microsoft.com/en-us/library/windows/desktop/dn280510(v=vs.85).aspx
         [DllImport("Shcore.dll")]
-        private static extern IntPtr GetDpiForMonitor([In]IntPtr hmonitor, [In]DpiType dpiType, [Out]out uint dpiX, [Out]out uint dpiY);
+        private static extern IntPtr GetDpiForMonitor([In] IntPtr hmonitor, [In] DpiType dpiType, [Out] out uint dpiX, [Out] out uint dpiY);
     }
 
     //https://msdn.microsoft.com/en-us/library/windows/desktop/dn280511(v=vs.85).aspx
